requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain-groq
streamlit
langchain-ollama